Fix SandS behavior
When typing, for example, "sudo apt", I often ended up typing the space
while pressing the "o" key. This resulted in "sudoapt". By increasing
the TAPPING_TERM, this problem is fixed. However, it introduces a new
problem where if I type "o" and "space" both within TAPPING_TERM it
would not be registered as "shift-o" but rather each character
individually. For this problem we use HOLD_ON_OTHER_KEY_PRESS.

https://docs.qmk.fm/tap_hold#hold-on-other-key-press