Change LALT and RALT to LAPO and RAPC (set to curly braces)
This change is made so that it would be easier to navigate between
paragraphs in Vim. I wasn't using the parentheses (original LAPO and
RAPC) because it was simply faster to just press Shift-9 or Shift-0.

On the other hand, I found that I use the curly braces `{` and `}` more
often when navigating between code blocks.